requirements.txt
Flask
dotenv
google-generativeai
opencv-python
mediapipe
dlib
scipy
numpy